requirements.txt
# --- Core Framework ---
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0

# --- LlamaIndex & OpenAI (Otak Baru) ---
llama-index-core==0.10.11
llama-index-llms-openai==0.1.6
llama-index-embeddings-openai==0.1.6
llama-index-vector-stores-chroma==0.1.4

# --- Database ---
chromadb==0.4.22

# --- Utilitas ---
pandas==2.1.3
pydantic==2.5.2